Added some sell chest stuff

+++ src/main/java/dk/banannus/generators/utils/isInt.java
@@ -3,7 +3,7 @@ package dk.banannus.generators.utils;
 public class isInt {
 	public static boolean isInt(String s) {
 		try {
-			Integer.parseInt(s);
+			Double.parseDouble(s);
 		} catch (NumberFormatException nfe) {
 			return false;
 		}
